Move 'utils::bytearray' to 'expr/bytearray'
